# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AllMart API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: placeOrder
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin/pm only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /checkout/quote:
    post:
      operationId: getCheckoutQuote
      tags: [orders]
      summary: Calculate the current checkout total including available discounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutQuoteRequest"
      responses:
        "200":
          description: Current checkout quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutQuote"
...
components:
  schemas:
...
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [placed, confirmed, dispatched, delivered, cancelled]
      required: [status]
...
    OrderItem:
      type: object
      properties:
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        imageUrl: { type: string }
      required: [productId, productName, quantity, unitPrice, imageUrl]
    Order:
      type: object
      properties:
        id: { type: integer }
        status: { type: string, enum: [placed, confirmed, dispatched, delivered, cancelled] }
        total: { type: number }
        currency: { type: string }
        trackingCode: { type: string }
        shippingAddress: { type: string }
        receiverName:
          type: string
          nullable: true
        receiverEmail:
          type: string
          nullable: true
        receiverPhone:
          type: string
          nullable: true
        paymentMethod:
          type: string
          enum: [stripe, bank_transfer, pay_on_delivery]
          nullable: true
        paymentScreenshotUrl:
          type: string
          nullable: true
        paymentNote:
          type: string
          nullable: true
        cashbackCode:
          type: string
          nullable: true
        cashbackDiscount:
          type: number
          nullable: true
        bonusDiscount:
          type: number
          nullable: true
        shippingFee: { type: number }
        placedBy: { type: string, enum: [user, ai] }
        createdAt: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
      required: [id, status, total, currency, trackingCode, shippingAddress, shippingFee, placedBy, createdAt, items]
    PlaceOrderRequest:
      type: object
      properties:
        shippingAddress: { type: string, minLength: 3 }
        receiverName: { type: string }
        receiverEmail: { type: string, format: email }
        receiverPhone: { type: string }
        cashbackCode: { type: string }
        paymentMethod:
          type: string
          enum: [stripe, bank_transfer, pay_on_delivery]
        paymentScreenshotUrl: { type: string }
        paymentNote: { type: string, maxLength: 2000 }
        bonusApplied: { type: boolean }
        expectedTotal: { type: number }
        expectedBonusDiscount: { type: number }
        placedBy:
          type: string
          enum: [user, ai]
      required: [shippingAddress]
    CheckoutQuoteRequest:
      type: object
      properties:
        cashbackCode: { type: string }
        bonusApplied: { type: boolean }
    CheckoutQuote:
      type: object
      properties:
        subtotal: { type: number }
        shippingFee: { type: number }
        cashbackDiscount: { type: number }
        bonusDiscount: { type: number }
        total: { type: number }
        currency: { type: string }
        bonusBalance: { type: number }
      required: [subtotal, shippingFee, cashbackDiscount, bonusDiscount, total, currency, bonusBalance]
...