# lib/api-zod/src/generated/types/checkoutQuote.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AllMart API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckoutQuote {
  subtotal: number;
  shippingFee: number;
  cashbackDiscount: number;
  bonusDiscount: number;
  total: number;
  currency: string;
  bonusBalance: number;
}