# lib/api-zod/src/generated/types/checkoutQuoteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AllMart API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckoutQuoteRequest {
  cashbackCode?: string;
  bonusApplied?: boolean;
}