# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AllMart API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderPaymentMethod } from "./orderPaymentMethod";
import type { OrderPlacedBy } from "./orderPlacedBy";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  status: OrderStatus;
  total: number;
  currency: string;
  trackingCode: string;
  shippingAddress: string;
  receiverName?: string | null;
  receiverEmail?: string | null;
  receiverPhone?: string | null;
  paymentMethod?: OrderPaymentMethod;
  paymentScreenshotUrl?: string | null;
  paymentNote?: string | null;
  cashbackCode?: string | null;
  cashbackDiscount?: number | null;
  bonusDiscount?: number | null;
  shippingFee: number;
  placedBy: OrderPlacedBy;
  createdAt: string;
  items: OrderItem[];
}