# lib/api-zod/src/generated/types/placeOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AllMart API
 * OpenAPI spec version: 0.1.0
 */
import type { PlaceOrderRequestPaymentMethod } from "./placeOrderRequestPaymentMethod";
import type { PlaceOrderRequestPlacedBy } from "./placeOrderRequestPlacedBy";

export interface PlaceOrderRequest {
  /** @minLength 3 */
  shippingAddress: string;
  receiverName?: string;
  receiverEmail?: string;
  receiverPhone?: string;
  cashbackCode?: string;
  paymentMethod?: PlaceOrderRequestPaymentMethod;
  paymentScreenshotUrl?: string;
  /** @maxLength 2000 */
  paymentNote?: string;
  bonusApplied?: boolean;
  expectedTotal?: number;
  expectedBonusDiscount?: number;
  placedBy?: PlaceOrderRequestPlacedBy;
}